Created new component of GroupController with buttons to operate a group + Integrated newly created component in Group component
