+++ frontend/src/Components/GroupController/GroupController.css
@@ -1,0 +1,7 @@
+.groupcontroller-container button {
+  margin: 0 1rem;
+  padding: 0 0.3rem;
+  border: 1px solid #004d73;
+  border-radius: 3px;
+  color: #004d73;
+}
