+++ frontend/src/Components/NewGroup/NewGroup.css
@@ -1,3 +1,7 @@
+.newgroup-container {
+  margin: 1rem 0;
+}
+
 .hidden {
   display: none;
 }
